backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.7.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cohere==4.21
colorama==0.4.6
cryptography==41.0.4
dataclasses-json==0.5.14
docx2txt==0.8
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.80.0
fastavro==1.8.2
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
langchain==0.0.263
langchain-experimental==0.0.8
langsmith==0.0.26
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.8
pyflakes==3.1.0
PyMuPDF==1.23.0
PyMuPDFb==1.23.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.19.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
